fix(svgr): named type without mixed import not work
