+++ e2e/cases/svg/svg-assets/index.test.ts
@@ -1,0 +1,27 @@
+import { expect, test } from '@playwright/test';
+import { build, gotoPage } from '@e2e/helper';
+
+test('SVGR basic usage', async ({ page }) => {
+  const rsbuild = await build({
+    cwd: __dirname,
+    runServer: true,
+  });
+
+  await gotoPage(page, rsbuild);
+
+  // test svg asset
+  await expect(
+    page.evaluate(
+      `document.getElementById('test-img').src.includes('static/svg/app')`,
+    ),
+  ).resolves.toBeTruthy();
+
+  // test svg asset in css
+  await expect(
+    page.evaluate(
+      `getComputedStyle(document.getElementById('test-css')).backgroundImage.includes('static/svg/app')`,
+    ),
+  ).resolves.toBeTruthy();
+
+  await rsbuild.close();
+});
