--- e2e/cases/svg/svg/src/App.css
@@ -1,5 +1,0 @@
-#test-css {
-  width: 50px;
-  height: 50px;
-  background-image: url('./app.svg');
-}
