+++ e2e/cases/svg/svgr-default-export-component/index.test.ts
@@ -1,0 +1,17 @@
+import { expect, test } from '@playwright/test';
+import { build, gotoPage } from '@e2e/helper';
+
+test('Use SVGR and default export React component', async ({ page }) => {
+  const rsbuild = await build({
+    cwd: __dirname,
+    runServer: true,
+  });
+
+  await gotoPage(page, rsbuild);
+
+  await expect(
+    page.evaluate(`document.getElementById('test-svg').tagName === 'svg'`),
+  ).resolves.toBeTruthy();
+
+  await rsbuild.close();
+});
