+++ e2e/cases/svg/svgr-default-export-component/src/App.jsx
