+++ e2e/cases/svg/svgr-external-react/src/index.js
