+++ e2e/cases/svg/svgr-external-react/static/index.html
