+++ e2e/cases/svg/svgr-named-export-component copy/rsbuild.config.ts
@@ -1,0 +1,13 @@
+import { pluginReact } from '@rsbuild/plugin-react';
+import { pluginSvgr } from '@rsbuild/plugin-svgr';
+
+export default {
+  plugins: [
+    pluginReact(),
+    pluginSvgr({
+      svgrOptions: {
+        exportType: 'named',
+      },
+    }),
+  ],
+};
