+++ e2e/cases/svg/svgr-named-export-component copy/src/App.jsx
@@ -1,0 +1,12 @@
+import { ReactComponent as Logo } from './app.svg';
+
+function App() {
+  return (
+    <div>
+      <div id="test">Hello Rsbuild!</div>
+      <Logo id="test-svg" />
+    </div>
+  );
+}
+
+export default App;
