+++ e2e/cases/svg/svgr-named-export-component copy/src/index.js
